utils: let format_utc_offset handle microseconds
format_utc_offset can be called with the return value of
g_date_time_get_utc_offset returning microseconds in a gint64. It is
also still called with the return value of icaltimezone_get_utc_offset
in seconds. Deal with both.

https://bugzilla.gnome.org/show_bug.cgi?id=762438

+++ src/gcal-utils.h
@@ -121,6 +121,6 @@ void            get_source_parent_name_color                    (GcalManager    
                                                                  gchar                **name,
                                                                  gchar                **color);
 
-gchar*          format_utc_offset                               (gint                   offset);
+gchar*          format_utc_offset                               (gint64                 offset);
 
 #endif // __GCAL_UTILS_H__
